Prepare a data.table of queries and variables for the FMAT.

Prepare a data.table of queries and variables for the FMAT.

Usage

FMAT_query(
  query = "Text with [MASK], optionally with {TARGET} and/or {ATTRIB}.",
  MASK = .(),
  TARGET = .(),
  ATTRIB = .(),
  unmask.id = 1
)

Arguments

query

Query text (should be a character string/vector with at least one [MASK] token). Multiple queries share the same set of MASK, TARGET, and ATTRIB. For multiple queries with different MASK, TARGET, and/or ATTRIB, please use FMAT_query_bind to combine them.

MASK

A named list of [MASK] target words. Must be single words in the vocabulary of a certain masked language model. For model vocabulary, see, e.g., https://huggingface.co/bert-base-uncased/raw/main/vocab.txt

Note that infrequent words may be not included in a model's vocabulary, and in this case you may insert the words into the context by specifying either TARGET or ATTRIB.

TARGET, ATTRIB

A named list of Target/Attribute words or phrases. If specified, then query must contain {TARGET} and/or {ATTRIB} (in all uppercase and in braces) to be replaced by the words/phrases.

unmask.id

If there are multiple [MASK] in query, this argument will be used to determine which one is to be unmasked. Defaults to the 1st [MASK].

Value

A data.table of queries and variables.

See also

Examples

FMAT_query("[MASK] is a nurse.", MASK = .(Male="He", Female="She"))
#>                 query   MASK M_pair M_word
#> 1: [MASK] is a nurse.   Male      1     He
#> 2: [MASK] is a nurse. Female      1    She

FMAT_query(
  c("[MASK] is {TARGET}.", "[MASK] works as {TARGET}."),
  MASK = .(Male="He", Female="She"),
  TARGET = .(Occupation=cc("a doctor, a nurse, an artist"))
)
#>     qid                     query   MASK M_pair M_word     TARGET       T_pair
#>  1:   1       [MASK] is {TARGET}.   Male      1     He Occupation Occupation.1
#>  2:   1       [MASK] is {TARGET}. Female      1    She Occupation Occupation.1
#>  3:   1       [MASK] is {TARGET}.   Male      1     He Occupation Occupation.2
#>  4:   1       [MASK] is {TARGET}. Female      1    She Occupation Occupation.2
#>  5:   1       [MASK] is {TARGET}.   Male      1     He Occupation Occupation.3
#>  6:   1       [MASK] is {TARGET}. Female      1    She Occupation Occupation.3
#>  7:   2 [MASK] works as {TARGET}.   Male      1     He Occupation Occupation.1
#>  8:   2 [MASK] works as {TARGET}. Female      1    She Occupation Occupation.1
#>  9:   2 [MASK] works as {TARGET}.   Male      1     He Occupation Occupation.2
#> 10:   2 [MASK] works as {TARGET}. Female      1    She Occupation Occupation.2
#> 11:   2 [MASK] works as {TARGET}.   Male      1     He Occupation Occupation.3
#> 12:   2 [MASK] works as {TARGET}. Female      1    She Occupation Occupation.3
#>        T_word
#>  1:  a doctor
#>  2:  a doctor
#>  3:   a nurse
#>  4:   a nurse
#>  5: an artist
#>  6: an artist
#>  7:  a doctor
#>  8:  a doctor
#>  9:   a nurse
#> 10:   a nurse
#> 11: an artist
#> 12: an artist

FMAT_query(
  "The [MASK] {ATTRIB}.",
  MASK = .(Male=cc("man, boy"),
           Female=cc("woman, girl")),
  ATTRIB = .(Masc=cc("is masculine, has a masculine personality"),
             Femi=cc("is feminine, has a feminine personality"))
)
#>                    query   MASK M_pair M_word ATTRIB      A_pair
#>  1: The [MASK] {ATTRIB}.   Male      1    man   Masc Masc-Femi.1
#>  2: The [MASK] {ATTRIB}.   Male      2    boy   Masc Masc-Femi.1
#>  3: The [MASK] {ATTRIB}. Female      1  woman   Masc Masc-Femi.1
#>  4: The [MASK] {ATTRIB}. Female      2   girl   Masc Masc-Femi.1
#>  5: The [MASK] {ATTRIB}.   Male      1    man   Masc Masc-Femi.2
#>  6: The [MASK] {ATTRIB}.   Male      2    boy   Masc Masc-Femi.2
#>  7: The [MASK] {ATTRIB}. Female      1  woman   Masc Masc-Femi.2
#>  8: The [MASK] {ATTRIB}. Female      2   girl   Masc Masc-Femi.2
#>  9: The [MASK] {ATTRIB}.   Male      1    man   Femi Masc-Femi.1
#> 10: The [MASK] {ATTRIB}.   Male      2    boy   Femi Masc-Femi.1
#> 11: The [MASK] {ATTRIB}. Female      1  woman   Femi Masc-Femi.1
#> 12: The [MASK] {ATTRIB}. Female      2   girl   Femi Masc-Femi.1
#> 13: The [MASK] {ATTRIB}.   Male      1    man   Femi Masc-Femi.2
#> 14: The [MASK] {ATTRIB}.   Male      2    boy   Femi Masc-Femi.2
#> 15: The [MASK] {ATTRIB}. Female      1  woman   Femi Masc-Femi.2
#> 16: The [MASK] {ATTRIB}. Female      2   girl   Femi Masc-Femi.2
#>                          A_word
#>  1:                is masculine
#>  2:                is masculine
#>  3:                is masculine
#>  4:                is masculine
#>  5: has a masculine personality
#>  6: has a masculine personality
#>  7: has a masculine personality
#>  8: has a masculine personality
#>  9:                 is feminine
#> 10:                 is feminine
#> 11:                 is feminine
#> 12:                 is feminine
#> 13:  has a feminine personality
#> 14:  has a feminine personality
#> 15:  has a feminine personality
#> 16:  has a feminine personality

FMAT_query(
  "The {TARGET} has a [MASK] association with {ATTRIB}.",
  MASK = .(H="high", L="low"),
  TARGET = .(Flower=cc("rose, iris, lily"),
             Insect=cc("ant, cockroach, spider")),
  ATTRIB = .(Pos=cc("health, happiness, love, peace"),
             Neg=cc("death, sickness, hatred, disaster"))
)
#>                                                    query MASK M_pair M_word
#>  1: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#>  2: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#>  3: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#>  4: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#>  5: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#>  6: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#>  7: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#>  8: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#>  9: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 10: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 11: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 12: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 13: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 14: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 15: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 16: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 17: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 18: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 19: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 20: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 21: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 22: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 23: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 24: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 25: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 26: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 27: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 28: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 29: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 30: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 31: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 32: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 33: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 34: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 35: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 36: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 37: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 38: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 39: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 40: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 41: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 42: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 43: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 44: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 45: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 46: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 47: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 48: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 49: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 50: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 51: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 52: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 53: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 54: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 55: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 56: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 57: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 58: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 59: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 60: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 61: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 62: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 63: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 64: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 65: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 66: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 67: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 68: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 69: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 70: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 71: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 72: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 73: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 74: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 75: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 76: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 77: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 78: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 79: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 80: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 81: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 82: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 83: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 84: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 85: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 86: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 87: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 88: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 89: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 90: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 91: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 92: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 93: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 94: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#> 95: The {TARGET} has a [MASK] association with {ATTRIB}.    H      1   high
#> 96: The {TARGET} has a [MASK] association with {ATTRIB}.    L      1    low
#>                                                    query MASK M_pair M_word
#>     TARGET    T_word ATTRIB    A_word
#>  1: Flower      rose    Pos    health
#>  2: Flower      rose    Pos    health
#>  3: Flower      iris    Pos    health
#>  4: Flower      iris    Pos    health
#>  5: Flower      lily    Pos    health
#>  6: Flower      lily    Pos    health
#>  7: Flower      rose    Pos happiness
#>  8: Flower      rose    Pos happiness
#>  9: Flower      iris    Pos happiness
#> 10: Flower      iris    Pos happiness
#> 11: Flower      lily    Pos happiness
#> 12: Flower      lily    Pos happiness
#> 13: Flower      rose    Pos      love
#> 14: Flower      rose    Pos      love
#> 15: Flower      iris    Pos      love
#> 16: Flower      iris    Pos      love
#> 17: Flower      lily    Pos      love
#> 18: Flower      lily    Pos      love
#> 19: Flower      rose    Pos     peace
#> 20: Flower      rose    Pos     peace
#> 21: Flower      iris    Pos     peace
#> 22: Flower      iris    Pos     peace
#> 23: Flower      lily    Pos     peace
#> 24: Flower      lily    Pos     peace
#> 25: Flower      rose    Neg     death
#> 26: Flower      rose    Neg     death
#> 27: Flower      iris    Neg     death
#> 28: Flower      iris    Neg     death
#> 29: Flower      lily    Neg     death
#> 30: Flower      lily    Neg     death
#> 31: Flower      rose    Neg  sickness
#> 32: Flower      rose    Neg  sickness
#> 33: Flower      iris    Neg  sickness
#> 34: Flower      iris    Neg  sickness
#> 35: Flower      lily    Neg  sickness
#> 36: Flower      lily    Neg  sickness
#> 37: Flower      rose    Neg    hatred
#> 38: Flower      rose    Neg    hatred
#> 39: Flower      iris    Neg    hatred
#> 40: Flower      iris    Neg    hatred
#> 41: Flower      lily    Neg    hatred
#> 42: Flower      lily    Neg    hatred
#> 43: Flower      rose    Neg  disaster
#> 44: Flower      rose    Neg  disaster
#> 45: Flower      iris    Neg  disaster
#> 46: Flower      iris    Neg  disaster
#> 47: Flower      lily    Neg  disaster
#> 48: Flower      lily    Neg  disaster
#> 49: Insect       ant    Pos    health
#> 50: Insect       ant    Pos    health
#> 51: Insect cockroach    Pos    health
#> 52: Insect cockroach    Pos    health
#> 53: Insect    spider    Pos    health
#> 54: Insect    spider    Pos    health
#> 55: Insect       ant    Pos happiness
#> 56: Insect       ant    Pos happiness
#> 57: Insect cockroach    Pos happiness
#> 58: Insect cockroach    Pos happiness
#> 59: Insect    spider    Pos happiness
#> 60: Insect    spider    Pos happiness
#> 61: Insect       ant    Pos      love
#> 62: Insect       ant    Pos      love
#> 63: Insect cockroach    Pos      love
#> 64: Insect cockroach    Pos      love
#> 65: Insect    spider    Pos      love
#> 66: Insect    spider    Pos      love
#> 67: Insect       ant    Pos     peace
#> 68: Insect       ant    Pos     peace
#> 69: Insect cockroach    Pos     peace
#> 70: Insect cockroach    Pos     peace
#> 71: Insect    spider    Pos     peace
#> 72: Insect    spider    Pos     peace
#> 73: Insect       ant    Neg     death
#> 74: Insect       ant    Neg     death
#> 75: Insect cockroach    Neg     death
#> 76: Insect cockroach    Neg     death
#> 77: Insect    spider    Neg     death
#> 78: Insect    spider    Neg     death
#> 79: Insect       ant    Neg  sickness
#> 80: Insect       ant    Neg  sickness
#> 81: Insect cockroach    Neg  sickness
#> 82: Insect cockroach    Neg  sickness
#> 83: Insect    spider    Neg  sickness
#> 84: Insect    spider    Neg  sickness
#> 85: Insect       ant    Neg    hatred
#> 86: Insect       ant    Neg    hatred
#> 87: Insect cockroach    Neg    hatred
#> 88: Insect cockroach    Neg    hatred
#> 89: Insect    spider    Neg    hatred
#> 90: Insect    spider    Neg    hatred
#> 91: Insect       ant    Neg  disaster
#> 92: Insect       ant    Neg  disaster
#> 93: Insect cockroach    Neg  disaster
#> 94: Insect cockroach    Neg  disaster
#> 95: Insect    spider    Neg  disaster
#> 96: Insect    spider    Neg  disaster
#>     TARGET    T_word ATTRIB    A_word