Run the mask filling pipeline on multiple models.

Run the mask filling pipeline on multiple models.

Usage

FMAT_run(
  models,
  data,
  file = NULL,
  progress = c(FALSE, TRUE, "none", "text", "time"),
  parallel = FALSE,
  ncores = min(length(models), parallel::detectCores()),
  warning = TRUE
)

Arguments

models

Language model(s):

Model names (usually the BERT-based models) at HuggingFace.

A list of mask filling pipelines loaded by FMAT_load. * You will need to rerun FMAT_load if you restart the R session.

data

A data.table returned from FMAT_query or FMAT_query_bind.

file

File name of .RData to save the returned data.

progress

Show a progress bar: "none" (FALSE), "text" (TRUE), "time".

parallel

Parallel processing. Defaults to FALSE. If TRUE, then models must be model names rather than from FMAT_load.

* For small-scale data, parallel processing would instead be slower because it takes time to create a parallel cluster.

ncores

Number of CPU cores to be used in parallel processing. Defaults to the minimum of the number of models and your CPU cores.

warning

Warning of out-of-vocabulary word(s). Defaults to TRUE.

Value

A data.table (of new class fmat) appending data

with these new variables:

model: model name.

output: complete sentence output with unmasked token.

token: actual token to be filled in the blank mask (a note "out-of-vocabulary" will be added if the original word is not found in the model vocabulary).

prop: (raw) conditional probability of the unmasked token given the provided context, estimated by the masked language model. * It is NOT SUGGESTED to directly interpret the raw probabilities because the contrast between a pair of probabilities is more interpretable. See summary.fmat.

Details

The function will also automatically adjust for the compatibility of tokens used in certain models: (1) for uncased models (e.g., ALBERT), it turns tokens to lowercase; (2) for models that use <mask> rather than [MASK], it automatically uses the corrected mask token; (3) for models that require a prefix to estimate whole words than subwords (e.g., ALBERT, RoBERTa), it adds a certain prefix (usually a white space; \u2581 for ALBERT and XLM-RoBERTa, \u0120 for RoBERTa and DistilRoBERTa).

Note that these changes only affect the token variable in the returned data, but will not affect the M_word variable.z Thus, users may analyze their data based on the unchanged M_word rather than the token.

See also

Examples

# Running the example requires the models downloaded
# You will need to rerun `FMAT_load` if you restart the R session
# \donttest{
models = FMAT_load(c("bert-base-uncased", "bert-base-cased"))
#> Initializing environment...
#> Using existing conda environment  textrpp_condaenv  for text installation
#> . 
#> text: torch==1.11.0, transformers==4.19.2, numpy, nltk will be installed.  Installing text required python packages...
#> 
#> ✔ Successfully installed Python modules in conda environment.
#> Error: No valid Python or conda environment.
#> 
#>        You may need to specify the version of Python:
#>          RStudio -> Tools -> Global/Project Options
#>          -> Python -> Select -> Conda Environments
#>          -> Choose ".../textrpp_condaenv/python.exe"

dq = FMAT_query(
  c("[MASK] is {TARGET}.", "[MASK] works as {TARGET}."),
  MASK = .(Male="He", Female="She"),
  TARGET = .(Occupation=cc("a doctor, a nurse, an artist"))
)
data1 = FMAT_run(models, dq)
#> Using existing conda environment  textrpp_condaenv  for text installation
#> . 
#> text: torch==1.11.0, transformers==4.19.2, numpy, nltk will be installed.  Installing text required python packages...
#> 
#> ✔ Successfully installed Python modules in conda environment.
#> Error: No valid Python or conda environment.
#> 
#>        You may need to specify the version of Python:
#>          RStudio -> Tools -> Global/Project Options
#>          -> Python -> Select -> Conda Environments
#>          -> Choose ".../textrpp_condaenv/python.exe"
summary(data1)
#> Error in eval(expr, envir, enclos): object 'data1' not found

data2 = FMAT_run(
  models,
  FMAT_query(
    "The {TARGET} has a [MASK] association with {ATTRIB}.",
    MASK = .(H="high", L="low"),
    TARGET = .(Flower=cc("rose, iris, lily"),
               Insect=cc("ant, cockroach, spider")),
    ATTRIB = .(Pos=cc("health, happiness, love, peace"),
               Neg=cc("death, sickness, hatred, disaster"))
  ))
#> Using existing conda environment  textrpp_condaenv  for text installation
#> . 
#> text: torch==1.11.0, transformers==4.19.2, numpy, nltk will be installed.  Installing text required python packages...
#> 
#> ✔ Successfully installed Python modules in conda environment.
#> Error: No valid Python or conda environment.
#> 
#>        You may need to specify the version of Python:
#>          RStudio -> Tools -> Global/Project Options
#>          -> Python -> Select -> Conda Environments
#>          -> Choose ".../textrpp_condaenv/python.exe"
summary(data2)
#> Error in eval(expr, envir, enclos): object 'data2' not found
# }